+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Graph widgets: Select node and hit `space` to toggle the attribute's value
 - Graph widgets: `ctrl +` to zoom in, `ctrl -` to zoom out and `ctrl 0` to reset zoom
 - Graph widgets: Allows moving graph with the arrow keys
+- Graph widgets: Inspector has now a button to allow centralizing the selection
 - New model plugin: Population growth `populationGrowth`
 - New model plugin: Conway's Game of Life cellular automaton `gameOfLife`
 - New model plugin: Elementary cellular automaton rules `cellularAutomata1D`